feat(Moonshot): add challenge 3
