+++ 11-Moonshot/Moonshot/Moonshot/Util/GridLayout.swift
@@ -1,0 +1,68 @@
+//  Date: 3/8/23
+//
+//  Author: Zai Santillan
+//  Github: @plskz
+
+
+import SwiftUI
+
+struct GridLayout: View {
+    let astronauts: [String: Astronaut]
+    let missions: [Mission]
+    
+    let columns = [
+        GridItem(.adaptive(minimum: 150))
+    ]
+    
+    var body: some View {
+        ScrollView {
+            LazyVGrid(columns: columns) {
+                ForEach(missions) { mission in
+                    NavigationLink {
+                        MissionView(mission: mission, astronauts: astronauts)
+                    } label: {
+                        VStack {
+                            Image(mission.image)
+                                .resizable()
+                                .scaledToFit()
+                                .frame(width: 100, height: 100)
+                                .padding()
+                            
+                            VStack {
+                                Text(mission.displayName)
+                                    .font(.headline)
+                                    .foregroundColor(.white)
+                                
+                                Text(mission.formattedShortLaunchDate)
+                                    .font(.caption)
+                                    .foregroundColor(.white.opacity(0.5))
+                                    .padding(.top, -5)
+                            }
+                            .padding(.vertical)
+                            .frame(maxWidth: .infinity)
+                            .background(.lightBackground)
+                        }
+                        .clipShape(RoundedRectangle(cornerRadius: 10))
+                        .overlay(
+                            RoundedRectangle(cornerRadius: 10)
+                                .stroke(.lightBackground)
+                        )
+                    }
+                }
+            }
+            .padding([.horizontal, .bottom])
+        }
+        .navigationTitle("Moonshot")
+        .background(.darkBackground)
+        .preferredColorScheme(.dark)
+    }
+}
+
+struct GridLayout_Previews: PreviewProvider {
+    static let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
+    static let missions: [Mission] = Bundle.main.decode("missions.json")
+    
+    static var previews: some View {
+        GridLayout(astronauts: astronauts, missions: missions)
+    }
+}
